Removing chips, Console errors, Objects in buttons, Cards as modals & Deviders
Aligned cards to the top.
Added in the remove functionality for the chips.
Changed the css that was causing the errors in the console.
Set buttons to allow objects to be passed as well as strings.
Set a card to be used as a modal, I think this is the way forward, to
remove the modal element and use cards.
Created decider component.
 - Added in indents & outdent