Updated ImageContainer(now Device)
Set up devices to allow for multiple props to be passed:
orientation='horizontal' device='iphone' variant="black" version="six"
Changed the way the styles work, They now pull from a json object that
stores all of the devices, versions and variants. This also allows for
defaults for all of these.