Added New MacBook back to devices
Re-added the new MacBook for the preview.
Note: When addeing more devices, keep the max dimension to 500px(will
use a scale variable to make bigger/small, maybe a width/height) and
set the sizes based upon the svgs sizes.
Also Viewport is the ‘Seems like’ screen size not the physical pixel
amount